change some requested chagnes
