+++ src/views/CategoryCreator/CategoryCreator.api.js
@@ -1,0 +1,17 @@
+import paths from 'constants/api';
+
+export const getBlogs = async () => {
+    const url = `${paths.BLOGS}`;
+
+    const headers = {
+        'Content-Type': 'application/json',
+    };
+
+    const response = await fetch(url, { headers, method: 'GET' });
+
+    if (response.status !== 200) {
+        throw new Error('błąd');
+    }
+
+    return response.json();
+}
